fix: move hmac to middleware and wrap docs to 80 characters
